[Progressing] create system spec about destroy action in members controller

+++ src/app/controllers/members_controller.rb
@@ -29,8 +29,10 @@ class MembersController < ApplicationController
 
   def destroy
     member = Member.find(params[:id])
-    flash[:success] = "#{member.user.name} removed"
-    member.destroy
+    unless member.owner
+      flash[:success] = "#{member.user.name} removed"
+      member.destroy 
+    end
   end
 
   private
